+++ includes/class-stancer-payment.php
@@ -119,7 +119,7 @@ class WC_Stancer_Payment extends WC_Stancer_Abstract_Table {
 	 * @param WC_Order $order Order to find.
 	 * @param array $payment_data Payment data used to create a new payment.
 	 * @param bool $generate_api_payment Do we need to generate a new payment if not already present.
-	 * @param string $status Status to find.
+	 * @param string[] $status Statuses to find.
 	 *
 	 * @return WC_Stancer_Payment
 	 */
@@ -127,17 +127,27 @@ class WC_Stancer_Payment extends WC_Stancer_Abstract_Table {
 		WC_Order $order,
 		array $payment_data = [],
 		bool $generate_api_payment = false,
-		string $status = 'pending'
+		array $status = []
 	) {
 		global $wpdb;
 
-		$row = $wpdb->get_row(
-			$wpdb->prepare(
-				"SELECT * FROM {$wpdb->prefix}wc_stancer_payment WHERE order_id = %d AND status = %s",
-				absint( $order->get_id() ),
-				esc_sql( $status )
-			)
-		);
+		// phpcs:disable WordPress.DB.PreparedSQL.NotPrepared
+
+		$values = [
+			absint( $order->get_id() ),
+		];
+		$sql = "SELECT * FROM {$wpdb->prefix}wc_stancer_payment WHERE order_id = %d";
+
+		if ( ! empty( $status ) ) {
+			$placeholder = implode( ', ', array_fill( 0, count( $status ), '%s' ) );
+			$sql .= ' AND status IN (' . $placeholder . ')';
+
+			array_push( $values, ...array_map( 'esc_sql', $status ) );
+		}
+
+		$row = $wpdb->get_row( $wpdb->prepare( $sql, $values ) );
+
+		// phpcs:enable
 
 		if ( ! $row && $generate_api_payment ) {
 			$api_payment = static::generate_api_payment( $order, $payment_data );
@@ -189,7 +199,7 @@ class WC_Stancer_Payment extends WC_Stancer_Abstract_Table {
 			$api_payment->return_url = $payment_data['return_url'];
 		}
 
-		WC_Stancer_Api::sent_object_to_api( $api_payment );
+		$api_payment->send();
 
 		return $api_payment;
 	}
@@ -206,7 +216,7 @@ class WC_Stancer_Payment extends WC_Stancer_Abstract_Table {
 	 * @return WC_Stancer_Payment
 	 */
 	public static function get_payment( $order, array $payment_data = [], bool $generate_api_payment = false ) {
-		$stancer_payment = static::find( $order, $payment_data, $generate_api_payment );
+		$stancer_payment = static::find( $order, $payment_data, $generate_api_payment, [ 'pending' ] );
 
 		return $stancer_payment;
 	}