+++ includes/class-stancer.php
@@ -66,35 +66,6 @@ class WC_Stancer {
 	}
 
 	/**
-	 * Initialize administration.
-	 *
-	 * @since 1.1.0
-	 */
-	public function init_admin() {
-		$this->install_database();
-
-		$options = get_option( 'woocommerce_stancer_settings' );
-		$replace = [
-			'description' => 'payment_description',
-			'title' => 'payment_option_text',
-		];
-		$updated = false;
-
-		foreach ( $replace as $key => $value ) {
-			if ( array_key_exists( $key, $options ) ) {
-				$options[ $value ] = $options[ $key ];
-				$updated = true;
-
-				unset( $options[ $key ] );
-			}
-		}
-
-		if ( $updated ) {
-			update_option( 'woocommerce_stancer_settings', $options );
-		}
-	}
-
-	/**
 	 * Create database at plugin activation.
 	 *
 	 * @since 1.0.0
@@ -102,13 +73,17 @@ class WC_Stancer {
 	public function install_database() {
 		global $wpdb;
 
-		$version = get_option( 'stancer', '1.0.0' );
+		$version = get_option( 'stancer-version', '0.0.0' );
+
+		if ( version_compare( STANCER_WC_VERSION, $version, '==' ) ) {
+			return;
+		}
+
+		require_once ABSPATH . 'wp-admin/includes/upgrade.php';
 
 		// phpcs:disable WordPress.DB.PreparedSQL.NotPrepared
 
-		if ( version_compare( '1.0.0', $version, '>=' ) ) {
-			require_once ABSPATH . 'wp-admin/includes/upgrade.php';
-
+		if ( version_compare( '1.0.0', $version, '>' ) ) {
 			$sql = 'CREATE TABLE ' . $wpdb->prefix . 'wc_stancer_card (
 				stancer_card_id int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT "Unique ID in this table",
 				user_id int(10) UNSIGNED NOT NULL COMMENT "User ID (see users table)",
@@ -166,6 +141,67 @@ class WC_Stancer {
 
 			dbDelta( $sql );
 		}
+
+		if ( version_compare( '1.1.0', $version, '>' ) ) {
+			$sql = 'CREATE TABLE ' . $wpdb->prefix . 'wc_stancer_subscription (
+				`stancer_subscription_id` int(10) UNSIGNED NOT NULL AUTO_INCREMENT COMMENT "Unique ID in this table",
+				`is_active` tinyint(1) UNSIGNED DEFAULT 1 COMMENT "Is still",
+				`subscription_id` int(10) UNSIGNED NOT NULL COMMENT "Subscription order ID (see posts table)",
+				`payment_id` binary(29) NOT NULL COMMENT "Source payment identifier",
+				`card_id` binary(29) NOT NULL COMMENT "ID of the card used",
+				`customer_id` binary(29) DEFAULT NULL COMMENT "ID of the customer used",
+				`datetime_created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT "Creation date and time",
+				`datetime_modified` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT "Last modification date and time",
+				PRIMARY KEY (`stancer_subscription_id`),
+				INDEX `subscription_id-is_active` (`subscription_id`, `is_active`)
+			);';
+
+			dbDelta( $sql );
+		}
+
+		// TODO: Remove it before release.
+		if ( version_compare( '1.1.0-rc3', $version, '=' ) ) {
+			$results = $wpdb->get_results(
+				'SELECT `stancer_subscription_id`, `subscription_id`, `payment_id` FROM `' . $wpdb->prefix . 'wc_stancer_subscription` ORDER BY 1 ASC;'
+			);
+
+			foreach ( $results as $result ) {
+				$customer_id = $wpdb->get_var(
+					$wpdb->prepare(
+						'SELECT `customer_id` FROM `' . $wpdb->prefix . 'wc_stancer_payment` WHERE `payment_id` = %s;',
+						$result->payment_id,
+					),
+				);
+
+				$subscriptions = $wpdb->get_results(
+					$wpdb->prepare(
+						'SELECT ID FROM `' . $wpdb->prefix . 'posts` WHERE `post_parent` = %d AND `post_type` = "shop_subscription";',
+						$result->subscription_id,
+					),
+				);
+
+				if ( $subscriptions ) {
+					foreach ( $subscriptions as $subscription ) {
+						$wpdb->update(
+							$wpdb->prefix . 'wc_stancer_subscription',
+							[
+								'customer_id' => $customer_id,
+								'subscription_id' => $subscription->ID,
+							],
+							[ 'stancer_subscription_id' => $result->stancer_subscription_id ],
+						);
+					}
+				} else {
+					$wpdb->update(
+						$wpdb->prefix . 'wc_stancer_subscription',
+						[ 'customer_id' => $customer_id ],
+						[ 'stancer_subscription_id' => $result->stancer_subscription_id ],
+					);
+				}
+			}
+		}
+
+		update_option( 'stancer-version', STANCER_WC_VERSION );
 	}
 
 	/**
@@ -174,8 +210,8 @@ class WC_Stancer {
 	 * @since 1.0.0
 	 */
 	private function load_actions() {
-		add_action( 'plugins_loaded', [ $this, 'load_gateway' ], 0 );
-		add_action( 'admin_init', [ $this, 'init_admin' ] );
+		add_action( 'plugins_loaded', [ $this, 'load_plugin' ] );
+		add_action( 'upgrader_process_complete', [ $this, 'upgrade_plugin' ] );
 		add_action( 'wc_ajax_create_order', [ $this, 'create_order' ] );
 		add_action( 'wp_enqueue_scripts', [ $this, 'load_public_hooks' ] );
 	}
@@ -216,9 +252,48 @@ class WC_Stancer {
 	}
 
 	/**
+	 * Load the plugin.
+	 *
+	 * @since 1.1.0
+	 */
+	public function load_plugin() {
+		$this->install_database();
+		$this->load_gateway();
+	}
+
+	/**
 	 * Fake run method.
 	 *
 	 * @since 1.0.0
 	 */
 	public function run() { }
+
+	/**
+	 * Upgrade the plugin.
+	 *
+	 * @since 1.1.0
+	 */
+	public function upgrade_plugin() {
+		$this->install_database();
+
+		$options = get_option( 'woocommerce_stancer_settings' );
+		$replace = [
+			'description' => 'payment_description',
+			'title' => 'payment_option_text',
+		];
+		$updated = false;
+
+		foreach ( $replace as $key => $value ) {
+			if ( array_key_exists( $key, $options ) ) {
+				$options[ $value ] = $options[ $key ];
+				$updated = true;
+
+				unset( $options[ $key ] );
+			}
+		}
+
+		if ( $updated ) {
+			update_option( 'woocommerce_stancer_settings', $options );
+		}
+	}
 }